Add files via upload
Reduce amount of platform-dependent code needed to run ShapeGen apps
