+++ windows-gdi/Makefile
@@ -4,9 +4,9 @@
 # Remember to run vcvars32.bat first to set up your build environment
 
 LIBFILES = user32.lib gdi32.lib Winmm.lib Msimg32.lib
-OBJFILES = winmain.obj textapp.obj bmpfile.obj gradient.obj pattern.obj \
+OBJFILES = winmain.obj textapp.obj bmpfile.obj pattern.obj gradient.obj renderer.obj\
            arc.obj curve.obj edge.obj path.obj stroke.obj thinline.obj
-BMPFILES = bricks.bmp fleur.bmp honeycomb.bmp star.bmp           
+BMPFILES = bricks.bmp fleur.bmp honeycomb.bmp star.bmp
 CC = cl.exe
 CDEBUG = -Zi
 LINK = link.exe
@@ -42,6 +42,9 @@ pattern.obj : pattern.cpp shapegen.h renderer.h
 gradient.obj : gradient.cpp shapegen.h renderer.h
 	$(CC) $(CDEBUG) -c gradient.cpp
 
+renderer.obj : renderer.cpp shapegen.h renderer.h
+	$(CC) $(CDEBUG) -c renderer.cpp
+
 arc.obj : arc.cpp shapegen.h shapepri.h
 	$(CC) $(CDEBUG) -c arc.cpp
 
@@ -61,72 +64,75 @@ thinline.obj : thinline.cpp shapegen.h shapepri.h
 	$(CC) $(CDEBUG) -c thinline.cpp
 
 demo.h : ..\demo.h
-        copy /y ..\demo.h
+	copy /y ..\demo.h
 
 shapegen.h : ..\shapegen.h
-        copy /y ..\shapegen.h
+	copy /y ..\shapegen.h
 
 shapepri.h : ..\shapepri.h
-        copy /y ..\shapepri.h
+	copy /y ..\shapepri.h
 
 renderer.h : ..\renderer.h
-        copy /y ..\renderer.h
+	copy /y ..\renderer.h
 
 nanosvg.h : ..\nanosvg.h
-        copy /y ..\nanosvg.h
+	copy /y ..\nanosvg.h
 
 arc.cpp : ..\arc.cpp
-        copy /y ..\arc.cpp
-        
+	copy /y ..\arc.cpp
+
 bmpfile.cpp : ..\bmpfile.cpp
-        copy /y ..\bmpfile.cpp
-        
+	copy /y ..\bmpfile.cpp
+
 curve.cpp : ..\curve.cpp
-        copy /y ..\curve.cpp
-        
+	copy /y ..\curve.cpp
+
 demo.cpp : ..\demo.cpp
-        copy /y ..\demo.cpp
-        
+	copy /y ..\demo.cpp
+
 edge.cpp : ..\edge.cpp
-        copy /y ..\edge.cpp
-        
+	copy /y ..\edge.cpp
+
 gradient.cpp : ..\gradient.cpp
-        copy /y ..\gradient.cpp
-        
+	copy /y ..\gradient.cpp
+
 pattern.cpp : ..\pattern.cpp
-        copy /y ..\pattern.cpp
-        
+	copy /y ..\pattern.cpp
+
 path.cpp : ..\path.cpp
-        copy /y ..\path.cpp
-        
+	copy /y ..\path.cpp
+
+renderer.cpp : ..\renderer.cpp
+	copy /y ..\renderer.cpp
+
 stroke.cpp : ..\stroke.cpp
-        copy /y ..\stroke.cpp
-        
+	copy /y ..\stroke.cpp
+
 svgview.cpp : ..\svgview.cpp
-        copy /y ..\svgview.cpp
-        
+	copy /y ..\svgview.cpp
+
 textapp.cpp : ..\textapp.cpp
-        copy /y ..\textapp.cpp
-        
+	copy /y ..\textapp.cpp
+
 thinline.cpp : ..\thinline.cpp
-        copy /y ..\thinline.cpp
-        
+	copy /y ..\thinline.cpp
+
 bricks.bmp : ..\bricks.bmp
-        copy /y ..\bricks.bmp
-        
+	copy /y ..\bricks.bmp
+
 fleur.bmp : ..\fleur.bmp
-        copy /y ..\fleur.bmp
-        
+	copy /y ..\fleur.bmp
+
 honeycomb.bmp : ..\honeycomb.bmp
-        copy /y ..\honeycomb.bmp
-        
+	copy /y ..\honeycomb.bmp
+
 star.bmp : ..\star.bmp
-        copy /y ..\star.bmp
-        
+	copy /y ..\star.bmp
+
 clean:
-        @del *.exe
-        @del *.ilk
-        @del *.obj
-        @del *.pdb
-        
+	@del *.exe
+	@del *.ilk
+	@del *.obj
+	@del *.pdb
+
